* commit 'origin/master': (88 commits)
  Don't say that a temporary WvString object in a const char argument list.
  Fix inconsequential typo in unit test
  insist on using a uint32 for the crc in adler32/crc32 in wvdigest.
  Replace a few printfs with wvcon->prints, helps remove compiler warnings
  Bump version to 4.5.1, in preparation for a release.
  Remove non-existence install-xplc target from install dependancies.
  Remove extern keyword from deserialize function.
  Remove Debian packaging from WvStreams. Packaging should be done downstream.
  Make use of valgrind optional.
  Search for qt3 using pkg-config (in addition to the other methods).
  Makes configure see --with-dbus as the same as the default.
  Unbreak finding of qt3 in configure.
  Apply some fixes from Fedora for compilation on gcc 4.3.
  Use the moc that we find in configure, not the first one in path.
  Apply fix to unbreak compilation by no longer setting obsolete sa_restorer flag
  wvtestrunner: catch segfaults in the subprogram.
  Oops, fix unit tests to match my recent dbus bool/byte format changes.
  wvdbusmsg: don't prepend 'y' or 'b' to bytes and booleans in get_str().
  Make the copy constructor for WvX509 much more efficient.
  get_signing_cert in ocsp response returns a reference to a cert, not a pointer
  ...

+++ include/wvdigest.h
@@ -137,7 +137,7 @@ private:
  */
 class WvCrc32Digest : public WvDigest
 {
-    unsigned long int crc;
+    uint32_t crc;
 
 public:
     WvCrc32Digest();
@@ -157,7 +157,7 @@ public:
  */
 class WvAdler32Digest : public WvDigest
 {
-    unsigned long int crc;
+    uint32_t crc;
 
 public:
     WvAdler32Digest();
